🔨 Remove install() targets
Instead, just set the output paths.

install is useful for libraries, but for executables that I want to be runnable from the build tree the same way as when 'installed', it's basically duplication.

Primary intention here though is to get rid of the `$<TARGET_PDB_FILE>` which isn't supported on the ninja generator